Renamed transformer to processor; create qualifiers for the different SbomProcessor implementations; removed SbomManipulator
