--- src/main/java/org/redhat/sbomer/transformer/SbomManipulator.java
@@ -1,50 +1,0 @@
-/**
- * JBoss, Home of Professional Open Source.
- * Copyright 2023 Red Hat, Inc., and individual contributors
- * as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.redhat.sbomer.transformer;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.enterprise.context.ApplicationScoped;
-
-import org.cyclonedx.model.Bom;
-
-@ApplicationScoped
-public class SbomManipulator {
-
-    private List<SbomTransformer> transformers = List.of();
-
-    public SbomManipulator addTransformer(SbomTransformer transformer) {
-        if (transformers.isEmpty()) {
-            transformers = new ArrayList<>();
-        }
-        transformers.add(transformer);
-        return this;
-    }
-
-    public Bom runTransformers(Bom bom) {
-        List<SbomTransformer> transformers = this.transformers;
-        if (!transformers.isEmpty()) {
-            for (SbomTransformer t : transformers) {
-                bom = t.transform(bom);
-            }
-        }
-        return bom;
-    }
-
-}
